fix: Phase 2 NaN — FP16 underflow/overflow in CGC, OT, Causal, logits
- CGC entropy: wrap in autocast(enabled=False), clamp(min=1e-6) for FP32
- OT Sinkhorn: cast to FP32 before log-domain computation
- Causal DAG reg: FP32 Taylor expansion (prevent FP16 overflow)
- _last_logits: .float() cast for stable loss computation
- GradScaler: try-except guard for assertion error on all-NaN epochs

+++ core/model/ple/model.py
@@ -1596,7 +1596,7 @@ class PLEModel(nn.Module):
                 if (_needs_logits
                         and tower is not None
                         and hasattr(tower, "_last_logits")):
-                    pred = tower._last_logits
+                    pred = tower._last_logits.float()
 
                 # Prepare inputs based on task type
                 if task_type in ("binary", "regression"):
